requirements.txt
Flask
googletrans==4.0.0-rc1
ffmpeg-python
whisper
pydub
